Add files via upload
Computer graphics is the process of making the design, 2D, 3D and animation of an object. Computer graphics can do many things, including modeling, simulation and visualization of an object or a problem. Modeling is a representation of how people describe or explain an object, system, or a concept, which is usually manifested by simplification or idealization. This can be represented by physical models (mockups, prototypes), the model image (design drawings, computer images), or mathematical formulae. OpenGL support this modeling capability as OpenGL has additional features to better produce something more realistic. OpenGL allows us to create a graph that can be run on any operating system with only minor adjustments. The 3-D graphics package designed here provides an interface for the users for handling the display and manipulation of 3D viewing model of a Building. Keyboard and mouse are the main input devices used. The aim of this project is to implement an animated bridge. The code contains the functions that accomplish tasks like changing color, opening and closing of the bridge as the cars pass by. This program nevertheless provides the basic functionality of smooth animation. The instructions are very simple to interpret and the interface is user friendly. Since this uses OpenGL platform, it becomes portable.